backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# AI/NLP
transformers==4.35.2
sentence-transformers==2.2.2
torch==2.1.1
spacy==3.7.2
openai==1.3.7
google-generativeai==0.3.1

# PDF processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0

# Database
motor==3.3.2
pymongo==4.6.0
beanie==1.23.6

# Authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Utilities
numpy==1.26.2
scikit-learn==1.3.2
pandas==2.1.3
aiofiles==23.2.1
